source-asana: update to estuary-cdk and canonical structure
