--- source-asana/source_asana/.dockerignore
@@ -1,6 +1,0 @@
-*
-!Dockerfile
-!main.py
-!source_asana
-!setup.py
-!secrets
