--- source-asana/source_asana/Dockerfile
@@ -1,16 +1,0 @@
-FROM python:3.9-slim
-
-# Bash is installed for more convenient debugging.
-RUN apt-get update && apt-get install -y bash && rm -rf /var/lib/apt/lists/*
-
-WORKDIR /airbyte/integration_code
-COPY source_asana ./source_asana
-COPY main.py ./
-COPY setup.py ./
-RUN pip install .
-
-ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
-ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
-
-LABEL io.airbyte.version=0.6.1
-LABEL io.airbyte.name=airbyte/source-asana
