--- source-asana/source_asana/acceptance-test-config.yml
@@ -1,42 +1,0 @@
-# See [Connector Acceptance Tests](https://docs.airbyte.com/connector-development/testing-connectors/connector-acceptance-tests-reference)
-# for more information about how to configure these tests
-connector_image: airbyte/source-asana:dev
-test_strictness_level: high
-acceptance_tests:
-  spec:
-    tests:
-      - spec_path: "source_asana/spec.json"
-  connection:
-    tests:
-      - config_path: "secrets/config.json"
-        status: "succeed"
-      - config_path: "integration_tests/invalid_config.json"
-        status: "failed"
-  discovery:
-    tests:
-      - config_path: "secrets/config.json"
-  basic_read:
-    tests:
-      - config_path: "secrets/config.json"
-        timeout_seconds: 7200
-        expect_records:
-          bypass_reason: "Bypassed until dedicated sandbox account is up and running. Please follow https://github.com/airbytehq/airbyte/issues/19662."
-        empty_streams:
-          - name: custom_fields
-            bypass_reason: "This stream is not available on the account we're currently using. Please follow https://github.com/airbytehq/airbyte/issues/19662."
-          - name: events
-            bypass_reason: "This stream is not available on our current account."
-          - name: portfolios_compact
-            bypass_reason: "This stream is not available on our current account, it requires a business/enterprise account"
-          - name: portfolios
-            bypass_reason: "This stream is not available on our current account, it requires a business/enterprise account"
-          - name: portfolios_memberships
-            bypass_reason: "This stream is not available on our current account, it requires a business/enterprise account"
-  full_refresh:
-    # tests:
-    #   - config_path: "secrets/config.json"
-    #     configured_catalog_path: "integration_tests/configured_catalog.json"
-    #     timeout_seconds: 7200
-    bypass_reason: "As we are using an internal account the data is not frozen and results of `two-sequential-reads` are flaky. Please follow https://github.com/airbytehq/airbyte/issues/19662."
-  incremental:
-    bypass_reason: "Incremental syncs are not supported on this connector."
