--- source-asana/source_asana/integration_tests/invalid_config.json
@@ -1,3 +1,0 @@
-{
-  "credentials": { "personal_access_token": "<wrong_access_token>" }
-}
