--- source-asana/source_asana/integration_tests/sample_config.json
@@ -1,3 +1,0 @@
-{
-  "access_token": "Your Personal Access Token."
-}
