--- source-asana/source_asana/main.py
@@ -1,13 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-import sys
-
-from airbyte_cdk.entrypoint import launch
-from source_asana import SourceAsana
-
-if __name__ == "__main__":
-    source = SourceAsana()
-    launch(source, sys.argv[1:])
