--- source-asana/source_asana/requirements.txt
@@ -1,1 +1,0 @@
--e .
