+++ source-asana/source_asana/schemas/stories_compact.json
