--- source-asana/source_asana/setup.py
@@ -1,25 +1,0 @@
-#
-# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
-#
-
-
-from setuptools import find_packages, setup
-
-MAIN_REQUIREMENTS = [
-    "airbyte-cdk~=0.2",
-]
-
-TEST_REQUIREMENTS = ["pytest-mock~=3.6.1", "pytest~=6.1", "requests-mock~=1.9.3"]
-
-setup(
-    name="source_asana",
-    description="Source implementation for Asana.",
-    author="Airbyte",
-    author_email="contact@airbyte.io",
-    packages=find_packages(),
-    install_requires=MAIN_REQUIREMENTS,
-    package_data={"": ["*.json", "schemas/*.json", "schemas/shared/*.json"]},
-    extras_require={
-        "tests": TEST_REQUIREMENTS,
-    },
-)
