+++ source-asana/tests/snapshots/source_asana_tests_test_snapshots__capture__capture.stdout.json
@@ -3,7 +3,8 @@
     "acmeCo/projects",
     {
       "_meta": {
-        "row_id": 1
+        "op": "u",
+        "row_id": 0
       },
       "archived": false,
       "color": "dark-teal",
@@ -61,7 +62,8 @@
     "acmeCo/sections",
     {
       "_meta": {
-        "row_id": 1
+        "op": "u",
+        "row_id": 0
       },
       "created_at": "2023-10-31T15:29:13.654Z",
       "gid": "1205846055668611",
@@ -77,7 +79,8 @@
     "acmeCo/sections",
     {
       "_meta": {
-        "row_id": 2
+        "op": "u",
+        "row_id": 1
       },
       "created_at": "2023-10-31T15:29:14.200Z",
       "gid": "1205846055668613",
@@ -93,7 +96,8 @@
     "acmeCo/sections",
     {
       "_meta": {
-        "row_id": 3
+        "op": "u",
+        "row_id": 2
       },
       "created_at": "2023-10-31T15:29:14.281Z",
       "gid": "1205846055668614",
@@ -109,7 +113,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 1
+        "op": "u",
+        "row_id": 0
       },
       "created_at": "2023-10-31T15:29:15.274Z",
       "created_by": {
@@ -129,7 +134,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 2
+        "op": "u",
+        "row_id": 1
       },
       "created_at": "2023-10-31T15:29:15.480Z",
       "created_by": {
@@ -149,7 +155,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 4
+        "op": "u",
+        "row_id": 3
       },
       "created_at": "2023-10-31T15:29:15.685Z",
       "created_by": {
@@ -169,7 +176,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 5
+        "op": "u",
+        "row_id": 4
       },
       "created_at": "2023-10-31T15:29:15.754Z",
       "created_by": {
@@ -189,7 +197,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 7
+        "op": "u",
+        "row_id": 6
       },
       "created_at": "2023-10-31T15:29:15.903Z",
       "created_by": {
@@ -209,7 +218,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 3
+        "op": "u",
+        "row_id": 2
       },
       "created_at": "2023-10-31T15:29:17.590Z",
       "created_by": {
@@ -229,7 +239,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 6
+        "op": "u",
+        "row_id": 5
       },
       "created_at": "2023-10-31T15:29:17.705Z",
       "created_by": {
@@ -249,7 +260,8 @@
     "acmeCo/stories",
     {
       "_meta": {
-        "row_id": 8
+        "op": "u",
+        "row_id": 7
       },
       "created_at": "2023-10-31T15:29:17.808Z",
       "created_by": {
@@ -269,7 +281,8 @@
     "acmeCo/tasks",
     {
       "_meta": {
-        "row_id": 1
+        "op": "u",
+        "row_id": 0
       },
       "assignee": {
         "gid": "1205846102972015"
@@ -329,7 +342,8 @@
     "acmeCo/tasks",
     {
       "_meta": {
-        "row_id": 2
+        "op": "u",
+        "row_id": 1
       },
       "assignee": {
         "gid": "1205846102972015"
@@ -389,7 +403,8 @@
     "acmeCo/tasks",
     {
       "_meta": {
-        "row_id": 3
+        "op": "u",
+        "row_id": 2
       },
       "assignee": null,
       "completed": false,
